v1.0.9

+++ src/structures/Channel/GroupChannel.js
@@ -12,14 +12,25 @@ module.exports = class GroupChannel extends TextBaseChannel {
         super(client,data)
 
         this.members = new GroupMemberManager(this)
+        this.invites = new GroupMemberManager(this)
 
         if (data) this._patch(data);
     }
 
-    get owner(){
-        return this.members.cache.get(this.extra.groupExtra.creator)
+    get me(){
+        return this.members.cache.get(this.client.user.id)
     }
-    
+
+    get owner(){
+        let member = this.members.cache.get(this.extra.groupExtra.creator)
+        if(!member) return this.client.users.add({id:this.extra.groupExtra.creator},false,{id: this.extra.groupExtra.creator})
+        else return member
+    }
+
+    get joined(){
+        return this.members.cache.has(this.client.user.id)
+    }
+
     _patch(data){
         super._patch(data)
         if('picturePath' in data) {
@@ -35,11 +46,21 @@ module.exports = class GroupChannel extends TextBaseChannel {
              */
             this.extra = data.extra;
             if('groupExtra' in data.extra) {
-                for (const id in data.extra.groupExtra.memberMids) {
-                    this.members && this.members.add({
-                        id: id,
-                        timestamp: data.extra.groupExtra.memberMids[id]
-                    })
+                if('memberMids' in data.extra.groupExtra) {
+                    for (const id in data.extra.groupExtra.memberMids) {
+                        this.members && this.members.add({
+                            id: id,
+                            timestamp: data.extra.groupExtra.memberMids[id]
+                        })
+                    }
+                }
+                if('inviteeMids' in data.extra.groupExtra) {
+                    for (const id in data.extra.groupExtra.inviteeMids) {
+                        this.invites && this.invites.add({
+                            id: id,
+                            timestamp: data.extra.groupExtra.inviteeMids[id]
+                        })
+                    }
                 }
             }
         }
@@ -49,13 +70,7 @@ module.exports = class GroupChannel extends TextBaseChannel {
         return CONSENT.line_server.CDN_PATH + this.picturePath
     }
     async fetch(){
-        let chat = (await this.client.api.getChats({
-            chatMids: [this.id],
-            withMembers: true,
-            withInvitees: true
-        })).chats
-        if(!chat[0]) return;
-        this._patch(chat[0])
+        return this.client.groups.fetch(this.id)
     }
     /*
     * leave this group
@@ -71,19 +86,49 @@ module.exports = class GroupChannel extends TextBaseChannel {
      */
     invite(users){
         if(!Array.isArray(users)) users = [users]
-        return this.client.api.inviteIntoChat({
-            reqSeq: 0,
-            id: this.id,
-            targetUserMids: users.map(user=>this.client.users.resolveID(user))
-        })
+        users = users.map(user=>this.client.users.resolveID(user)).filter(id => !this.members.cache.has(id) && !this.invites.cache.has(id))
+        if(users.length >= 1) {
+            return this.client.api.inviteIntoChat({
+                reqSeq: 0,
+                id: this.id,
+                targetUserMids: users
+            })
+        } else return false
     }
 
-    kick(users){
+    kick(users = []){
         if(!Array.isArray(users)) users = [users];
-        return this.client.api.deleteOtherFromChat({
+        users = users.map(user=>this.client.users.resolveID(user)).filter(id => id != this.client.user.id && id != this.owner.id)
+        if(users.length >= 1) {
+            return this.client.api.deleteOtherFromChat({
+                reqSeq: 0,
+                chatMid: this.id,
+                targetUserMids: users
+            })
+        } else return false
+    }
+    
+    /**
+     * Reject This Group Invite If not Joined
+     * @return {Promise<Object>}
+     */
+     async accept(){
+        if(!this.invites.cache.has(this.client.user.id)) return false;
+        return this.client.api.acceptChatInvitation({
             reqSeq: 0,
-            chatMid: this.id,
-            targetUserMids: users.map(user=>this.client.users.resolveID(user))
+            chatMid: this.id
+        })
+    }
+    
+    /**
+     * Reject This Group Invite If not Joined
+     * @return {Promise<Object>}
+     */
+    async reject(){
+        if(!this.invites.cache.has(this.client.user.id)) return false;
+        return await this.client.api.rejectChatInvitation({
+            reqSeq: 0,
+            chatMid: this.id
         })
     }
 }
